@@ -1,98 +1,83 @@
 {
   "name": "@zondax/ledger-cosmos-js",
-  "author": "Zondax AG",
-  "license": "Apache-2.0",
   "version": "4.0.0",
   "description": "Node API for Cosmos App (Ledger Nano S/S+/X)",
-  "main": "./dist/index.js",
-  "typings": "./dist/index.d.ts",
-  "types": "./dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/cosmos/ledger-cosmos-js.git"
-  },
   "keywords": [
     "Ledger",
     "Typescript",
     "Tendermint",
     "Cosmos"
   ],
+  "homepage": "https://github.com/Zondax/ledger-cosmos-js",
   "bugs": {
-    "url": "https://github.com/cosmos/ledger-cosmos-js/issues"
+    "url": "https://github.com/Zondax/ledger-cosmos-js/issues"
   },
-  "homepage": "https://github.com/cosmos/ledger-cosmos-js",
-  "resolutions": {
-    "@ledgerhq/hw-transport": "6.27.1",
-    "@ledgerhq/devices": "6.27.1",
-    "@ledgerhq/hw-transport-node-hid": "6.27.1",
-    "@ledgerhq/hw-transport-node-hid-noevents": "6.27.1"
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cosmos/ledger-cosmos-js.git"
+  },
+  "license": "Apache-2.0",
+  "author": "Zondax AG",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "typings": "./dist/index.d.ts",
+  "files": [
+    "dist/**",
+    "LICENSE",
+    "package.json"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "format": "FORCE_COLOR=1 prettier --write . && sort-package-json",
+    "format:check": "FORCE_COLOR=1 prettier --check .",
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "test": "yarn build && jest .",
+    "upgrade": "bunx npm-check-updates -i"
   },
   "dependencies": {
-    "@ledgerhq/hw-transport": "6.30.6",
-    "@zondax/ledger-js": "^0.10.0",
-    "bech32": "^2.0.0",
+    "@ledgerhq/hw-transport": "6.31.4",
+    "@noble/hashes": "^1.7.1",
+    "@scure/base": "^1.2.4",
+    "@zondax/ledger-js": "^1.2.0",
     "buffer": "^6.0.3",
-    "crypto": "^1.0.1",
-    "crypto-browserify": "^3.12.0",
-    "ripemd160": "^2.0.2",
-    "stream-browserify": "^3.0.0"
+    "ripemd160": "^2.0.2"
   },
   "devDependencies": {
-    "@ledgerhq/hw-transport-node-hid": "6.28.6",
-    "@ledgerhq/hw-transport-webusb": "6.28.6",
-    "@ledgerhq/hw-transport-mocker": "^6.28.6",
-    "@swc/core": "^1.6.1",
-    "@types/jest": "^29.5.12",
-    "@types/ledgerhq__hw-transport": "4.21.8",
-    "@types/node": "^20.14.2",
-    "@typescript-eslint/eslint-plugin": "^7.13.0",
-    "@typescript-eslint/parser": "^7.13.0",
-    "bip32": "4.0.0",
+    "@ledgerhq/hw-transport-mocker": "^6.29.4",
+    "@ledgerhq/hw-transport-node-hid": "^6.29.5",
+    "@noble/curves": "^1.8.1",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.2",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.14.0",
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "bip32": "5.0.0-rc.0",
     "bip39": "3.1.0",
-    "copyfiles": "^2.4.1",
-    "core-js": "^3.37.1",
-    "crypto-js": "4.2.0",
-    "eslint": "^9.5.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.6.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-unused-imports": "^3.2.0",
-    "eslint-plugin-tsdoc": "^0.2.17",
+    "eslint": "^9.24.0",
+    "eslint-config-prettier": "^10.1.2",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "eslint-plugin-tsdoc": "^0.4.0",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "jest": "^29.7.0",
-    "jest-serial-runner": "^1.2.0",
-    "prettier": "^3.3.2",
-    "secp256k1": "^5.0.0",
-    "ts-jest": "29.1.5",
+    "jest-runner": "^29.7.0",
+    "jest-serial-runner": "^1.2.1",
+    "prettier": "^3.5.3",
+    "sort-package-json": "^3.0.0",
+    "ts-jest": "29.3.1",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "webpack": "^5.92.0"
+    "typescript": "^5.8.3"
+  },
+  "volta": {
+    "node": "20.11.1",
+    "bun": "1.0.20"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "moduleDirectories": [
     "node_modules",
     "dist"
-  ],
-  "browserslist": [
-    "> 1%",
-    "last 2 versions"
-  ],
-  "files": [
-    "dist/*",
-    "LICENSE",
-    "yarn.lock"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "peerDependencies": {
-    "@babel/core": "^7.17.5",
-    "@types/node": "^17.0.21"
-  },
-  "scripts": {
-    "build": "tsc",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "format": "prettier --write .",
-    "test": "yarn build && jest"
-  }
+  ]
 }
