@@ -1,4 +1,7 @@
 module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-};
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  transformIgnorePatterns: ['^.+\\.js$'],
+  modulePaths: ['<rootDir>/src', '<rootDir>/tests'],
+  runner: 'jest-serial-runner',
+}
