@@ -1,5 +1,5 @@
 /** ******************************************************************************
- *  (c) 2018 - 2023 Zondax AG
+ *  (c) Zondax AG
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  ******************************************************************************* */
-export const CLA = 0x55;
+export const CLA = 0x55
 
 export const enum P1_VALUES {
   ONLY_RETRIEVE = 0x00,
@@ -25,4 +25,4 @@ export const enum P2_VALUES {
   TEXTUAL = 0x1,
 }
 
-export const PKLEN = 33;
+export const PKLEN = 33
