@@ -1,13 +1,13 @@
 export interface ResponseAddress {
-  bech32_address: string;
-  compressed_pk: Buffer;
+  bech32_address: string
+  compressed_pk: Buffer
 }
 
 export interface ResponsePubkey {
-  bech32_address: string;
-  compressed_pk: Buffer;
+  bech32_address: string
+  compressed_pk: Buffer
 }
 
 export interface ResponseSign {
-  signature: Buffer;
+  signature: Buffer
 }
